refactor: extract allowProjectOwnerAsAssignee logic
